backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
chromadb==0.5.5
pydantic==2.9.0
python-dotenv==1.0.1
python-docx==1.1.0
pypdf==5.0.0
requests==2.32.3
